Move selection/creation of custom file to top of init.
I set a custom file where all package-specific customizations are placed.  These are
settings that are altered using the customization engine.  Since I prefer to put all
my customizations in the init file(s), everything in this file is auto-generated by
one package or another.  When these changes are in the init.el file, it makes it very
hard to manage it with version control.

Move this block of code from the bottom of the init file to the top as part of the
core setup.