Comment out `cc-mode' use-package stanza.
This approach is not working, even though I modeled it on JWiegley's init.el file.  There are a
couple of issues in this version which are corrected on another host, but have not yet been
merged.  However, this whole use-package stanza may go away in favor of using custom hook
functions to configure c- and c++-mode.