Re-wrote init file from the ground up.
Init file has been blown away and replaced with a complete re-write.
There is also a helper file in the lisp directory which aids in setup.
Most packages I use on a daily basis have been added to this new
config.  There is still some config to do and packages to add, but the
bulk of the work has been done and this config file is in a very good
place.

init.el
lisp/mw-basic-setup.el