Add some packages and move some ':config' calls to `use-package' to ':init'
Add the following packages:
1. toml-mode
2. yaml-mode
3. cql-mode
4. cmake-mode
5. systemd

Move some dependent use-package calls into ':init' section of
enclosing use-package expression.  This ensures that the dependent
use-package exression is evaluated regardless of the deferral status
of the enclosing expression.  The ':config' block may be deferred if
'defer' is set to a non-nil value.

init.el