Update ace-jump-mode bindings to be more responsive.
Previously we had the ace-jump commands bound to keychords.  This did not work well
because there was a pause introduced by processing the keychord, and if the chord
didn't register then the whole process would simply insert the keychord characters
into the buffer.  These types of functions are not good candidates for use with
keychords.  Instead these functions have been bound to the "C-c a" prefix.

init.el