Alter Jinx config so it only loads in non-programming buffers.
Use the `hack-local-variables-hook' to run a lambda which will enable Jinx only if the
major mode is not derived from `prog-mode'.  `hack-local-variables-hook' is a normal
hook that is run after a file's local variables have been processed.  This guarantees
it is run after the major mode has been enabled for the buffer.

Add a short comments section at the end of the init.el file to explain this and other
points of interest.

init.el